tools/pretraining_data_builder/requirements.txt
httpx>=0.27.2
python-dotenv>=1.0.0
orjson>=3.9.10
rich>=13.7.0
click>=8.1.7
msgspec>=0.18.4
asyncclick>=8.1.3.4
numpy
gdal
pyproj
mercantile
Pillow
shapely
imageio
geopandas
pyresample
lmdb